requirements.txt
MarkupSafe~=3.0.2
filelock~=3.18.0
fsspec~=2025.3.2
jinja2~=3.1.6
mpmath~=1.3.0
networkx~=3.4.2
sympy~=1.13.1
torch~=2.6.0
torchvision~=0.21.0
pillow~=11.1.0
opencv-python~=4.11.0.86
numpy~=2.2.4
scipy~=1.15.2